+++ package.json
@@ -62,7 +62,6 @@
         "@tauri-apps/plugin-dialog": "2.0.0-beta.8",
         "@tauri-apps/plugin-fs": "2.0.0-beta.8",
         "@tauri-apps/plugin-http": "2.0.0-beta.8",
-        "@tauri-apps/plugin-shell": "2.0.0-beta.8",
         "@tauri-apps/plugin-store": "2.0.0-beta.8",
         "bits-ui": "0.21.12",
         "clsx": "2.1.1",
