+++ src-tauri/Cargo.toml
@@ -18,6 +18,5 @@ tauri = { version = "=2.0.0-beta.24", features = ["macos-private-api"] }
 tauri-plugin-dialog = { version = "=2.0.0-beta.10" }
 tauri-plugin-fs = { version = "=2.0.0-beta.11" }
 tauri-plugin-http = { version = "=2.0.0-beta.10" }
-tauri-plugin-shell = { version = "=2.0.0-beta.9" }
 tauri-plugin-store = { version = "=2.0.0-beta.10" }
 toml = { version = "=0.8.17" }
