--- src-tauri/src/workspace/mod.rs
@@ -1,2 +1,0 @@
-mod workspace;
-pub use workspace::*;
