+++ src-tauri/tauri.conf.json
@@ -15,9 +15,10 @@
                 "height": 755,
                 "minWidth": 800,
                 "minHeight": 512,
-                "decorations": false,
-                "transparent": true,
-                "shadow": false
+                "visible": false,
+                "decorations": true,
+                "titleBarStyle": "Overlay",
+                "hiddenTitle": true
             }
         ],
         "macOSPrivateApi": true,
